+++ superset-frontend/src/dashboard/components/nativeFilters/FilterBar/FilterSets/index.tsx
@@ -40,8 +40,7 @@ import { findExistingFilterSet } from './utils';
 import { useFilters, useNativeFiltersDataMask, useFilterSets } from '../state';
 import Footer from './Footer';
 import FilterSetUnit from './FilterSetUnit';
-import { getFilterBarTestId } from '../utils';
-import { TabIds } from '../utils';
+import { getFilterBarTestId, TabIds } from '../utils';
 
 const FilterSetsWrapper = styled.div`
   display: grid;
