+++ superset-frontend/src/dashboard/components/nativeFilters/FilterBar/Vertical.tsx
@@ -27,58 +27,19 @@ import React, {
   useRef,
   createContext,
 } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
 import cx from 'classnames';
-import {
-  DataMaskStateWithId,
-  DataMaskWithId,
-  Filter,
-  DataMask,
-  HandlerFunction,
-  styled,
-  t,
-  SLOW_DEBOUNCE,
-  isNativeFilter,
-} from '@superset-ui/core';
+import { HandlerFunction, styled, t, isNativeFilter } from '@superset-ui/core';
 import Icons from 'src/components/Icons';
 import { AntdTabs } from 'src/components';
-import { useHistory } from 'react-router-dom';
-import { usePrevious } from 'src/hooks/usePrevious';
 import { FeatureFlag, isFeatureEnabled } from 'src/featureFlags';
-import { updateDataMask, clearDataMask } from 'src/dataMask/actions';
-import { useImmer } from 'use-immer';
-import { isEmpty, isEqual, debounce } from 'lodash';
 import Loading from 'src/components/Loading';
-import { getInitialDataMask } from 'src/dataMask/reducer';
-import { URL_PARAMS } from 'src/constants';
-import { getUrlParam } from 'src/utils/urlUtils';
 import { EmptyStateSmall } from 'src/components/EmptyState';
-import { useTabId } from 'src/hooks/useTabId';
-import { RootState } from 'src/dashboard/types';
-import { checkIsApplyDisabled, TabIds } from './utils';
+import { TabIds, getFilterBarTestId, VerticalBarProps } from './utils';
 import FilterSets from './FilterSets';
-import {
-  useNativeFiltersDataMask,
-  useFilters,
-  useFilterSets,
-  useFilterUpdates,
-  useInitialization,
-} from './state';
-import { createFilterKey, updateFilterKey } from './keyValue';
+import { useFilterSets } from './state';
 import EditSection from './FilterSets/EditSection';
 import Header from './Header';
 import FilterControls from './FilterControls/FilterControls';
-import ActionButtons from './ActionButtons';
-import { getFilterBarTestId } from './utils';
-
-export interface VerticalFiltersBarProps {
-  filtersOpen: boolean;
-  toggleFiltersBar: any;
-  directPathToChild?: string[];
-  width: number;
-  height: number | string;
-  offset: number;
-}
 
 const BarWrapper = styled.div<{ width: number }>`
   width: ${({ theme }) => theme.gridUnit * 8}px;
@@ -162,192 +123,29 @@ const FilterBarEmptyStateContainer = styled.div`
   margin-top: ${({ theme }) => theme.gridUnit * 8}px;
 `;
 
-const EXCLUDED_URL_PARAMS: string[] = [
-  URL_PARAMS.nativeFilters.name,
-  URL_PARAMS.permalinkKey.name,
-];
-
-const publishDataMask = debounce(
-  async (
-    history,
-    dashboardId,
-    updateKey,
-    dataMaskSelected: DataMaskStateWithId,
-    tabId,
-  ) => {
-    const { location } = history;
-    const { search } = location;
-    const previousParams = new URLSearchParams(search);
-    const newParams = new URLSearchParams();
-    let dataMaskKey: string | null;
-    previousParams.forEach((value, key) => {
-      if (!EXCLUDED_URL_PARAMS.includes(key)) {
-        newParams.append(key, value);
-      }
-    });
-
-    const nativeFiltersCacheKey = getUrlParam(URL_PARAMS.nativeFiltersKey);
-    const dataMask = JSON.stringify(dataMaskSelected);
-    if (
-      updateKey &&
-      nativeFiltersCacheKey &&
-      (await updateFilterKey(
-        dashboardId,
-        dataMask,
-        nativeFiltersCacheKey,
-        tabId,
-      ))
-    ) {
-      dataMaskKey = nativeFiltersCacheKey;
-    } else {
-      dataMaskKey = await createFilterKey(dashboardId, dataMask, tabId);
-    }
-    if (dataMaskKey) {
-      newParams.set(URL_PARAMS.nativeFiltersKey.name, dataMaskKey);
-    }
-
-    // pathname could be updated somewhere else through window.history
-    // keep react router history in sync with window history
-    // replace params only when current page is /superset/dashboard
-    // this prevents a race condition between updating filters and navigating to Explore
-    if (window.location.pathname.includes('/superset/dashboard')) {
-      history.location.pathname = window.location.pathname;
-      history.replace({
-        search: newParams.toString(),
-      });
-    }
-  },
-  SLOW_DEBOUNCE,
-);
-
 export const FilterBarScrollContext = createContext(false);
-const VerticalFilterBar: React.FC<VerticalFiltersBarProps> = ({
-  filtersOpen,
-  toggleFiltersBar,
+const VerticalFilterBar: React.FC<VerticalBarProps> = ({
+  actions,
+  canEdit,
+  dataMaskSelected,
   directPathToChild,
-  width,
+  filtersOpen,
+  filterValues,
   height,
+  isDisabled,
+  isLoading,
   offset,
+  onSelectionChange,
+  toggleFiltersBar,
+  width,
 }) => {
-  const history = useHistory();
-  const dataMaskApplied: DataMaskStateWithId = useNativeFiltersDataMask();
   const [editFilterSetId, setEditFilterSetId] = useState<number | null>(null);
-  const [dataMaskSelected, setDataMaskSelected] =
-    useImmer<DataMaskStateWithId>(dataMaskApplied);
-  const dispatch = useDispatch();
-  const [updateKey, setUpdateKey] = useState(0);
-  const tabId = useTabId();
   const filterSets = useFilterSets();
   const filterSetFilterValues = Object.values(filterSets);
   const [tab, setTab] = useState(TabIds.AllFilters);
-  const filters = useFilters();
-  const previousFilters = usePrevious(filters);
-  const filterValues = Object.values(filters);
   const nativeFilterValues = filterValues.filter(isNativeFilter);
-  const dashboardId = useSelector<any, string>(
-    ({ dashboardInfo }) => dashboardInfo?.id,
-  );
-  const previousDashboardId = usePrevious(dashboardId);
-  const canEdit = useSelector<RootState, boolean>(
-    ({ dashboardInfo }) => dashboardInfo.dash_edit_perm,
-  );
-
   const [isScrolling, setIsScrolling] = useState(false);
   const timeout = useRef<any>();
-
-  const handleFilterSelectionChange = useCallback(
-    (
-      filter: Pick<Filter, 'id'> & Partial<Filter>,
-      dataMask: Partial<DataMask>,
-    ) => {
-      setDataMaskSelected(draft => {
-        // force instant updating on initialization for filters with `requiredFirst` is true or instant filters
-        if (
-          // filterState.value === undefined - means that value not initialized
-          dataMask.filterState?.value !== undefined &&
-          dataMaskSelected[filter.id]?.filterState?.value === undefined &&
-          filter.requiredFirst
-        ) {
-          dispatch(updateDataMask(filter.id, dataMask));
-        }
-
-        draft[filter.id] = {
-          ...(getInitialDataMask(filter.id) as DataMaskWithId),
-          ...dataMask,
-        };
-      });
-    },
-    [dataMaskSelected, dispatch, setDataMaskSelected],
-  );
-
-  useEffect(() => {
-    if (previousFilters && dashboardId === previousDashboardId) {
-      const updates = {};
-      Object.values(filters).forEach(currentFilter => {
-        const previousFilter = previousFilters?.[currentFilter.id];
-        if (!previousFilter) {
-          return;
-        }
-        const currentType = currentFilter.filterType;
-        const currentTargets = currentFilter.targets;
-        const currentDataMask = currentFilter.defaultDataMask;
-        const previousType = previousFilter?.filterType;
-        const previousTargets = previousFilter?.targets;
-        const previousDataMask = previousFilter?.defaultDataMask;
-        const typeChanged = currentType !== previousType;
-        const targetsChanged = !isEqual(currentTargets, previousTargets);
-        const dataMaskChanged = !isEqual(currentDataMask, previousDataMask);
-
-        if (typeChanged || targetsChanged || dataMaskChanged) {
-          updates[currentFilter.id] = getInitialDataMask(currentFilter.id);
-        }
-      });
-
-      if (!isEmpty(updates)) {
-        setDataMaskSelected(draft => ({ ...draft, ...updates }));
-        Object.keys(updates).forEach(key => dispatch(clearDataMask(key)));
-      }
-    }
-  }, [
-    JSON.stringify(filters),
-    JSON.stringify(previousFilters),
-    previousDashboardId,
-  ]);
-
-  const dataMaskAppliedText = JSON.stringify(dataMaskApplied);
-
-  useEffect(() => {
-    setDataMaskSelected(() => dataMaskApplied);
-  }, [dataMaskAppliedText, setDataMaskSelected]);
-
-  useEffect(() => {
-    publishDataMask(history, dashboardId, updateKey, dataMaskApplied, tabId);
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [dashboardId, dataMaskAppliedText, history, updateKey, tabId]);
-
-  const handleApply = useCallback(() => {
-    const filterIds = Object.keys(dataMaskSelected);
-    setUpdateKey(1);
-    filterIds.forEach(filterId => {
-      if (dataMaskSelected[filterId]) {
-        dispatch(updateDataMask(filterId, dataMaskSelected[filterId]));
-      }
-    });
-  }, [dataMaskSelected, dispatch]);
-
-  const handleClearAll = useCallback(() => {
-    const filterIds = Object.keys(dataMaskSelected);
-    filterIds.forEach(filterId => {
-      if (dataMaskSelected[filterId]) {
-        dispatch(clearDataMask(filterId));
-        setDataMaskSelected(draft => {
-          if (draft[filterId].filterState?.value !== undefined) {
-            draft[filterId].filterState!.value = undefined;
-          }
-        });
-      }
-    });
-  }, [dataMaskSelected, dispatch, setDataMaskSelected]);
 
   const openFiltersBar = useCallback(
     () => toggleFiltersBar(true),
@@ -372,13 +170,6 @@ const VerticalFilterBar: React.FC<VerticalFiltersBarProps> = ({
     };
   }, [onScroll]);
 
-  useFilterUpdates(dataMaskSelected, setDataMaskSelected);
-  const isApplyDisabled = checkIsApplyDisabled(
-    dataMaskSelected,
-    dataMaskApplied,
-    nativeFilterValues,
-  );
-  const isInitialized = useInitialization();
   const tabPaneStyle = useMemo(
     () => ({ overflow: 'auto', height, overscrollBehavior: 'contain' }),
     [height],
@@ -410,7 +201,7 @@ const VerticalFilterBar: React.FC<VerticalFiltersBarProps> = ({
         </CollapsedBar>
         <Bar className={cx({ open: filtersOpen })} width={width}>
           <Header toggleFiltersBar={toggleFiltersBar} />
-          {!isInitialized ? (
+          {!isLoading ? (
             <div css={{ height }}>
               <Loading />
             </div>
@@ -431,7 +222,7 @@ const VerticalFilterBar: React.FC<VerticalFiltersBarProps> = ({
                 {editFilterSetId && (
                   <EditSection
                     dataMaskSelected={dataMaskSelected}
-                    disabled={!isApplyDisabled}
+                    disabled={!isDisabled}
                     onCancel={() => setEditFilterSetId(null)}
                     filterSetId={editFilterSetId}
                   />
@@ -453,7 +244,7 @@ const VerticalFilterBar: React.FC<VerticalFiltersBarProps> = ({
                   <FilterControls
                     dataMaskSelected={dataMaskSelected}
                     directPathToChild={directPathToChild}
-                    onFilterSelectionChange={handleFilterSelectionChange}
+                    onFilterSelectionChange={onSelectionChange}
                   />
                 )}
               </AntdTabs.TabPane>
@@ -467,10 +258,10 @@ const VerticalFilterBar: React.FC<VerticalFiltersBarProps> = ({
               >
                 <FilterSets
                   onEditFilterSet={setEditFilterSetId}
-                  disabled={!isApplyDisabled}
+                  disabled={!isDisabled}
                   dataMaskSelected={dataMaskSelected}
                   tab={tab}
-                  onFilterSelectionChange={handleFilterSelectionChange}
+                  onFilterSelectionChange={onSelectionChange}
                 />
               </AntdTabs.TabPane>
             </StyledTabs>
@@ -493,19 +284,12 @@ const VerticalFilterBar: React.FC<VerticalFiltersBarProps> = ({
                 <FilterControls
                   dataMaskSelected={dataMaskSelected}
                   directPathToChild={directPathToChild}
-                  onFilterSelectionChange={handleFilterSelectionChange}
+                  onFilterSelectionChange={onSelectionChange}
                 />
               )}
             </div>
           )}
-          <ActionButtons
-            width={width}
-            onApply={handleApply}
-            onClearAll={handleClearAll}
-            dataMaskSelected={dataMaskSelected}
-            dataMaskApplied={dataMaskApplied}
-            isApplyDisabled={isApplyDisabled}
-          />
+          {actions}
         </Bar>
       </BarWrapper>
     </FilterBarScrollContext.Provider>
